+++ api/v1/nifidataflow_types.go
@@ -59,6 +59,10 @@ type UpdateRequest struct {
 	PercentCompleted int32 `json:"percentCompleted"`
 	// the state of the request
 	State string `json:"state"`
+	// whether or not this request was found.
+	NotFound bool `json:"notFound,omitempty"`
+	// the number of consecutive retries made in case of a NotFound error (limit: 3).
+	NotFoundRetryCount int32 `json:"notFoundRetryCount,omitempty"`
 }
 
 type DropRequest struct {
@@ -96,6 +100,10 @@ type DropRequest struct {
 	Dropped string `json:"dropped"`
 	// the state of the request
 	State string `json:"state"`
+	// whether or not this request was found.
+	NotFound bool `json:"notFound,omitempty"`
+	// the number of consecutive retries made in case of a NotFound error (limit: 3).
+	NotFoundRetryCount int32 `json:"notFoundRetryCount,omitempty"`
 }
 
 // NifiDataflowStatus defines the observed state of NifiDataflow
@@ -144,24 +152,15 @@ func (d *NifiDataflowSpec) GetSyncMode() DataflowSyncMode {
 }
 
 func (d *NifiDataflowSpec) SyncOnce() bool {
-	if d.GetSyncMode() == SyncOnce {
-		return true
-	}
-	return false
+	return d.GetSyncMode() == SyncOnce
 }
 
 func (d *NifiDataflowSpec) SyncAlways() bool {
-	if d.GetSyncMode() == SyncAlways {
-		return true
-	}
-	return false
+	return d.GetSyncMode() == SyncAlways
 }
 
 func (d *NifiDataflowSpec) SyncNever() bool {
-	if d.GetSyncMode() == SyncNever {
-		return true
-	}
-	return false
+	return d.GetSyncMode() == SyncNever
 }
 
 func (d *NifiDataflowSpec) GetParentProcessGroupID(rootProcessGroupId string) string {