+++ api/v1alpha1/nificluster_conversion.go
@@ -3,7 +3,7 @@ package v1alpha1
 import (
 	"fmt"
 
-	"github.com/konpyutaika/nifikop/api/v1"
+	v1 "github.com/konpyutaika/nifikop/api/v1"
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 )
 
@@ -67,15 +67,15 @@ func convertNifiClusterSpec(src *NifiClusterSpec, dst *v1.NifiCluster) error {
 	dst.Spec.OneNifiNodePerNode = src.OneNifiNodePerNode
 	dst.Spec.PropagateLabels = src.PropagateLabels
 	if src.NodeUserIdentityTemplate != nil {
-		dst.Spec.NodeUserIdentityTemplate = &(*src.NodeUserIdentityTemplate)
+		dst.Spec.NodeUserIdentityTemplate = src.NodeUserIdentityTemplate
 	}
 	dst.Spec.SidecarConfigs = src.SidecarConfigs
 	dst.Spec.TopologySpreadConstraints = src.TopologySpreadConstraints
 	if src.NifiControllerTemplate != nil {
-		dst.Spec.NifiControllerTemplate = &(*src.NifiControllerTemplate)
+		dst.Spec.NifiControllerTemplate = src.NifiControllerTemplate
 	}
 	if src.ControllerUserIdentity != nil {
-		dst.Spec.ControllerUserIdentity = &(*src.ControllerUserIdentity)
+		dst.Spec.ControllerUserIdentity = src.ControllerUserIdentity
 	}
 
 	convertNifiClusterSecretRef(src.SecretRef, dst)
@@ -100,9 +100,11 @@ func convertNifiClusterSecretRef(src SecretReference, dst *v1.NifiCluster) {
 
 func convertNifiClusterPodPolicy(src PodPolicy, dst *v1.NifiCluster) {
 	dst.Spec.Pod = v1.PodPolicy{
-		HostAliases: src.HostAliases,
-		Annotations: src.Annotations,
-		Labels:      src.Labels,
+		HostAliases:    src.HostAliases,
+		Annotations:    src.Annotations,
+		Labels:         src.Labels,
+		ReadinessProbe: nil,
+		LivenessProbe:  nil,
 	}
 }
 
@@ -356,6 +358,7 @@ func convertInternalListeners(src []InternalListenerConfig) []v1.InternalListene
 func convertSSLSecrets(src *SSLSecrets, dst *v1.ListenersConfig) {
 	if src == nil {
 		dst.SSLSecrets = nil
+		return
 	}
 	dst.SSLSecrets = &v1.SSLSecrets{
 		TLSSecretName: src.TLSSecretName,
@@ -504,15 +507,15 @@ func convertNifiClusterFromSpec(src *v1.NifiClusterSpec, dst *NifiCluster) error
 	dst.Spec.OneNifiNodePerNode = src.OneNifiNodePerNode
 	dst.Spec.PropagateLabels = src.PropagateLabels
 	if src.NodeUserIdentityTemplate != nil {
-		dst.Spec.NodeUserIdentityTemplate = &(*src.NodeUserIdentityTemplate)
+		dst.Spec.NodeUserIdentityTemplate = src.NodeUserIdentityTemplate
 	}
 	dst.Spec.SidecarConfigs = src.SidecarConfigs
 	dst.Spec.TopologySpreadConstraints = src.TopologySpreadConstraints
 	if src.NifiControllerTemplate != nil {
-		dst.Spec.NifiControllerTemplate = &(*src.NifiControllerTemplate)
+		dst.Spec.NifiControllerTemplate = src.NifiControllerTemplate
 	}
 	if src.ControllerUserIdentity != nil {
-		dst.Spec.ControllerUserIdentity = &(*src.ControllerUserIdentity)
+		dst.Spec.ControllerUserIdentity = src.ControllerUserIdentity
 	}
 
 	convertNifiClusterFromSecretRef(src.SecretRef, dst)
@@ -793,6 +796,7 @@ func convertFromInternalListeners(src []v1.InternalListenerConfig) []InternalLis
 func convertFromSSLSecrets(src *v1.SSLSecrets, dst *ListenersConfig) {
 	if src == nil {
 		dst.SSLSecrets = nil
+		return
 	}
 	dst.SSLSecrets = &SSLSecrets{
 		TLSSecretName: src.TLSSecretName,