+++ api/v1alpha1/nifiparametercontext_webhook.go
@@ -18,11 +18,7 @@ package v1alpha1
 
 import (
 	ctrl "sigs.k8s.io/controller-runtime"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
-
-// log is for logging in this package.
-var nifiparametercontextlog = logf.Log.WithName("nifiparametercontext-resource")
 
 func (r *NifiParameterContext) SetupWebhookWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewWebhookManagedBy(mgr).