+++ pkg/nificlient/client.go
@@ -173,7 +173,7 @@ func (n *nifiClient) Build() error {
 	n.client = n.newClient(config)
 
 	n.nodeClient = make(map[int32]*nigoapi.APIClient)
-	for nodeId, _ := range n.opts.NodesURI {
+	for nodeId := range n.opts.NodesURI {
 		nodeConfig := n.getNiNodeGoApiConfig(nodeId)
 		n.nodeClient[nodeId] = n.newClient(nodeConfig)
 	}
@@ -216,7 +216,6 @@ func (n *nifiClient) getNifiGoApiConfig() (config *nigoapi.Configuration) {
 	if n.opts.UseSSL {
 		transport = &http.Transport{}
 		config.Scheme = "HTTPS"
-		n.opts.TLSConfig.BuildNameToCertificate()
 		transport.TLSClientConfig = n.opts.TLSConfig
 		protocol = "https"
 	}
@@ -239,7 +238,7 @@ func (n *nifiClient) getNifiGoApiConfig() (config *nigoapi.Configuration) {
 	config.BasePath = fmt.Sprintf("%s://%s/nifi-api", protocol, n.opts.NifiURI)
 	config.Host = n.opts.NifiURI
 	if len(n.opts.NifiURI) == 0 {
-		for nodeId, _ := range n.opts.NodesURI {
+		for nodeId := range n.opts.NodesURI {
 			config.BasePath = fmt.Sprintf("%s://%s/nifi-api", protocol, n.opts.NodesURI[nodeId].RequestHost)
 			config.Host = n.opts.NodesURI[nodeId].RequestHost
 		}
@@ -256,7 +255,6 @@ func (n *nifiClient) getNiNodeGoApiConfig(nodeId int32) (config *nigoapi.Configu
 	if n.opts.UseSSL {
 		transport = &http.Transport{}
 		config.Scheme = "HTTPS"
-		n.opts.TLSConfig.BuildNameToCertificate()
 		transport.TLSClientConfig = n.opts.TLSConfig
 		protocol = "https"
 	}